// console/src/generated/sdk/ConnectionProfileWrite.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ConnectionProfileWrite = { id: string | null, name: string, backend: string | null, emulation: unknown | null, gzip: boolean | null, brotli: boolean | null, deflate: boolean | null, zstd: boolean | null, redirectMaxHops: number | null, retry: string | null, connectTimeoutMs: number | null, poolIdleTimeoutMs: number | null, poolMaxIdlePerHost: number | null, };